+++ prover/proof_producer/zkevm_rpcd_producer.go
@@ -70,9 +70,9 @@ type ProtocolInstance struct {
 	Prover                  string `json:"prover"`
 	GasUsed                 uint64 `json:"gas_used"`
 	ParentGasUsed           uint64 `json:"parent_gas_used"`
-	BlockMaxGasLimit        uint64 `json:"block_max_gas_limit"`
-	MaxTransactionsPerBlock uint64 `json:"max_transactions_per_block"`
-	MaxBytesPerTxList       uint64 `json:"max_bytes_per_tx_list"`
+	BlockMaxGasLimit        uint32 `json:"block_max_gas_limit"`
+	MaxTransactionsPerBlock uint32 `json:"max_transactions_per_block"`
+	MaxBytesPerTxList       uint32 `json:"max_bytes_per_tx_list"`
 }
 
 // RequestProofBodyResponse represents the JSON body of the response of the proof requests.
@@ -230,7 +230,7 @@ func (p *ZkevmRpcdProducer) requestProof(opts *ProofRequestOptions) (*RpcdOutput
 				Graffiti:                opts.Graffiti,
 				GasUsed:                 opts.GasUsed,
 				ParentGasUsed:           opts.ParentGasUsed,
-				BlockMaxGasLimit:        uint64(p.ProtocolConfig.BlockMaxGasLimit),
+				BlockMaxGasLimit:        p.ProtocolConfig.BlockMaxGasLimit,
 				MaxTransactionsPerBlock: p.ProtocolConfig.BlockMaxTransactions,
 				MaxBytesPerTxList:       p.ProtocolConfig.BlockMaxTxListBytes,
 			},
